requirements.txt
# Backend API dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI/ML - OpenAI
openai>=1.0.0
tiktoken>=0.12.0,<0.13.0

# AI/ML - Anthropic/Claude
anthropic>=0.71.0,<0.72.0

# Utilities
pydub>=0.25.1,<0.26.0
markitdown[docx,pdf,pptx,xls,xlsx,az-doc-intel]>=0.1.4,<0.2.0  # Document parsing with Office format support
pymupdf>=1.25.0  # Fast PDF text extraction
python-dotenv>=1.1.1,<2.0.0
